Migrate AttributeGroupResource to split-class layout
AttributeGroupResource  179 -> 70 lines  (+AttributeGroupForm 95, +AttributeGroupTable 55)

ResourceExtensionTest updated:
  - AttributeGroup removed from the legacy 'extendTable on Resource::class'
    dataset since it's no longer on the legacy shape.
  - New test case verifies the v5 pattern: registering against
    AttributeGroupTable::class with configureTable, asserting the column
    appears on the rendered ListAttributeGroups page. Same contract
    add-on authors will use going forward.

Closes the Catalog family.
